### 2019-01-05 #1155 @whedon
[PRE REVIEW]: LightTwinSVM: A Simple and Fast Implementation of Standard Twin Support Vector Machine Classifier
**Submitting author:** @mir-am (<a href="http://orcid.org/0000-0002-9818-7752">A. Mir</a>)
**Repository:** <a href="https://github.com/mir-am/LightTwinSVM" target ="_blank">https://github.com/mir-am/LightTwinSVM</a>
**Version:** v0.5.0
