### 2019-01-05 #1155 @mir-am
Hello
First, we are very grateful for the consideration of our paper.
Before starting the review process, we need to make some minor revisions to our paper to solve problems like an undefined reference.
However, we need an editor to be assigned. (We have suggested @arfon, Mr. Smith)

Regards,
Mir
