### 2019-01-05 #1155 @arfon
:wave: @mir-am - thanks for your submission to JOSS. Please go ahead and make the changes you need to for your paper and ask @whedon to recompile your paper here with the command `whedon generate pdf`.

> However, we need an editor to be assigned. (We have suggested @arfon, Mr. Smith)

I probably won't edit this submission - it may take another few days for an editor to be assigned.
