### 2019-01-06 e2df176
Minor revision to the paper of LightTwinSVM for Joss. [skip ci]
