+++ paper/paper.md
@@ -20,36 +20,44 @@ date:
 bibliography: paper.bib
 ---
 
+# Abstract
+This paper presents the `LightTwinSVM` program and its features. It is a simple and fast implementation of twin support vector machine  algorithm (TwinSVM). Numerical experiments on benchmark datasets shows the effectiveness of the `LightTwinSVM` program in terms of accuracy. This program can be used by researchers, students, and practitioners to solve classifications tasks.
+
 # Introduction
 
-Classification is a widely-used learning method in machine learning. The task of classification consists of training samples for which the class labels are available. On the basis of such training samples, a classifier learns a rule for predicting an unseen sample [@shalev2014]. To do a classification task, many algorithms have been proposed in machine learning literature such as Artificial Neural Network (ANN), Support Vector Machine (SVM), K-nearest neighbors (KNN) and Decision Trees. Among these classification algorithms, SVM classifier has relatively better prediction accuracy and generalization [@kotsiantis2007]. The main idea of SVM is to find the optimal separating hyperplane with the largest margin between the two classes. Figure below shows the geometric interpretation of SVM classifier.
+Classification is a widely-used learning method in machine learning. The task of classification consists of training samples for which the class labels are available. On the basis of such training samples, a classifier learns a rule for predicting an unseen sample [@shalev2014]. To do a classification task, many algorithms have been proposed in machine learning literature such as Artificial Neural Network (ANN), Support Vector Machine (SVM), K-nearest neighbors (KNN), and Decision Trees. Among these classification algorithms, SVM classifier has relatively better prediction accuracy and generalization [@kotsiantis2007]. The main idea of SVM is to find the optimal separating hyperplane with the largest margin between the two classes. Figure 1 shows the geometric illustration of SVM classifier.
 
-![Geometric interpretation of SVM classifier.](figs/SVM.jpg)
+![Geometric illustration of SVM classifier.](figs/SVM.png)
 
-Over the past decade, researchers have proposed new classifiers based on the SVM [@nayak2015]. Of these extensions of SVM, twin support vector machine (TwinSVM) [@jayadeva2007] has received more attention from scholars in the field of SVM research. This may be due to the novel idea of TwinSVM which is doing classification using two non-parallel hyperplanes. Each of which is as close as possible to samples of its own class and far from samples of the other class. To show the central idea of TwinSVM graphically, figure below shows the geometric interpretation of TwinSVM classifier.
+Over the past decade, researchers have proposed new classifiers based on the SVM [@nayak2015]. Of these extensions of SVM, twin support vector machine (TwinSVM) [@jayadeva2007] has received more attention from scholars in the field of SVM research. This may be due to the novel idea of TwinSVM which is doing classification using two non-parallel hyperplanes. Each of which is as close as possible to samples of its own class and far from samples of the other class. To show the central idea of TwinSVM graphically, Figure 2 shows the geometric interpretation of TwinSVM classifier.
 
-![Geometric interpretation of TwinSVM classifier.](figs/TwinSVM.jpg)
+![Geometric illustration of TwinSVM classifier.](figs/TwinSVM.png)
 
-SVM has several stable software packages and implementations such as `LIBSVM`[@chang2011libsvm] and `LIBLINEAR`[@fan2008liblinear]. These packages were used to implement SVM in `Scikit-learn` [@pedregosa2011scikit] which is a widely-used machine learning package for Python programming language. To solve a classification problem with SVM algorithm, one can use `Scikit-learn` with several lines of code in Python.
[…]
