+++ paper/paper.bib
@@ -101,3 +101,25 @@
    year = {2006},
    type = {Journal Article}
 }
+
+@article{nayak2015,
+   author = {Nayak, Janmenjoy and Naik, Bighnaraj and Behera, H},
+   title = {A comprehensive survey on support vector machine in data mining tasks: applications and challenges},
+   journal = {International Journal of Database Theory and Application},
+   volume = {8},
+   number = {1},
+   pages = {169-186},
+   ISSN = {2005-4270},
+   year = {2015},
+   type = {Journal Article}
+}
+
+@article{fan2008liblinear,
+  title={LIBLINEAR: A library for large linear classification},
+  author={Fan, Rong-En and Chang, Kai-Wei and Hsieh, Cho-Jui and Wang, Xiang-Rui and Lin, Chih-Jen},
+  journal={Journal of machine learning research},
+  volume={9},
+  number={Aug},
+  pages={1871--1874},
+  year={2008}
+}
