### 2019-01-06 #1155 @cMadan
@arfon, it seems that whedon updated the first comment with me as editor, but didn't 'acknowledge' the command or assign me to the issue in Github. Has anything changed with how it's supposed to do that...?
