### 2019-01-06 #1155 @cMadan
@mir-am, this should do it:
https://pandoc.org/MANUAL.html#tables
