### 2019-01-06 ad66002
A caption is added to the table. Also, footnotes. [skip ci]
