+++ paper/paper.md
@@ -29,7 +29,7 @@ Classification is a widely-used learning method in machine learning. The task of
 
 ![Geometric illustration of SVM classifier.](figs/SVM.png)
 
-Over the past decade, researchers have proposed new classifiers based on the SVM [@nayak2015]. Of these extensions of SVM, twin support vector machine (TwinSVM) [@jayadeva2007] has received more attention from scholars in the field of SVM research. This may be due to the novel idea of TwinSVM which is doing classification using two non-parallel hyperplanes. Each of which is as close as possible to samples of its own class and far from samples of the other class. To show the central idea of TwinSVM graphically, Figure 2 shows the geometric interpretation of TwinSVM classifier.
+Over the past decade, researchers have proposed new classifiers based on the SVM [@nayak2015]. Of these extensions of SVM, twin support vector machine (TwinSVM) [@jayadeva2007] has received more attention from scholars in the field of SVM research. This may be due to the novel idea of TwinSVM which is doing classification using two non-parallel hyperplanes. Each of which is as close as possible to samples of its own class and far from samples of the other class. To show the central idea of TwinSVM graphically, Figure 2 shows the geometric illustration of TwinSVM classifier.
 
 ![Geometric illustration of TwinSVM classifier.](figs/TwinSVM.png)
 
@@ -39,11 +39,11 @@ Even though TwinSVM is a popular classification algorithm in the field of SVM re
 
 # LightTwinSVM
 
-`LightTwinSVM` program is a simple and fast implementation of TwinSVM classifier. It is mostly written in Python and its main design goals are simplicity and speed. Also, this program is free, open source, and licensed under the terms of [GNU GPL v3](https://opensource.org/licenses/GPL-3.0). `LightTwinSVM` is built on top of `NumPy` [@walt2011], `Scikit-learn` [@pedregosa2011scikit], and `pandas` [@mckinney2011pandas].
+`LightTwinSVM` program is a simple and fast implementation of TwinSVM classifier. It is mostly written in Python and its main design goals are simplicity and speed. Also, this program is free, open source, and licensed under the terms of GNU GPL v3^[https://opensource.org/licenses/GPL-3.0]. `LightTwinSVM` is built on top of `NumPy` [@walt2011], `Scikit-learn` [@pedregosa2011scikit], and `pandas` [@mckinney2011pandas].
 
 `LightTwinSVM` program can be used by both researchers in the field of SVM research and by students in courses on pattern recognition and machine learning. Moreover, this software can be applied to a wide variety of research applications such as text classification, image or video recognition, medical diagnosis, and bioinformatics. For example, `LightTwinSVM` was used for the numerical experiments in our previous research paper [@mir2018]. 
 
-The main features of `LightTwinSVM` are the following:
+The main features of the `LightTwinSVM` program are the following:
 
 - To make its usage simple, a **command-line application** was created to help users solve classification tasks step-by-step.
 - Since speed is one of the design goals, the **clipDCD optimization algorithm** [@peng2014] is employed which is a simple and fast external optimizer. It was improved and implemented in C++.
@@ -55,9 +55,10 @@ The main features of `LightTwinSVM` are the following:
 - **CSV** and **LIBSVM** formats are supported for importing datasets.
 - Detailed classification results are saved in a spreadsheet file so that results can be analyzed and interpreted.
 
[…]
