### 2019-01-06 ba78911
Fixed table and figures references in text. [skip ci]
