### 2019-01-06 #1155 @mir-am
> @mir-am, this should do it:
> https://pandoc.org/MANUAL.html#tables

Thanks.
We revised our paper.
We would like to start the review process.
