### 2019-01-08 #1155 @arfon
> @arfon, not sure what happened earlier, but didn't seem like @whedon acted properly earlier. This time it worked though.

Weird. Not sure. Seems like a minor hiccup...
