### 2019-01-14 #1155 @mir-am
> @mir-am, I am asking around for additional reviewers and hopefully someone will volunteer.
>
> I was looking at the figures in the paper, are the SVM and TwinSVM illustrations based on the same data?

Thanks.
The dataset for both illustrations is nearly the same. Data points are randomly selected from an artificial dataset. However, we can add new illustrations based on the exact same dataset.
