### 2019-01-17 #1155 @rougier
I can review but it'll take me some time (3/4 weeks) because I'm a bit busy right now.
