### 2019-01-17 #1155 @cMadan
@mir-am, I think it would be better if they are the exact same, thanks!

@rougier, that sounds fine to me. I won't start the review yet in case someone else is also able/interested to review, but if you could comment here when your free, I'll start the review even if no one else has volunteered.
