### 2019-01-19 #1155 @mir-am
> PDF failed to compile for issue #1155 with the following error:
>
> % Total % Received % Xferd Average Speed Time Time Time Current
> Dload Upload Total Spent Left Speed
>
> 0 0 0 0 0 0 0 0 --:--:-- --:--:-- --:--:-- 0
> 100 15 0 15 0 0 234 0 --:--:-- --:--:-- --:--:-- 238
> Error producing PDF.
> ! Undefined control sequence.
> \SQSPL@scan ->\futurelet \SQSPL@next
> \SQSPL@scani
> l.388 \texttt{Scikit-learn}'s SVM\label{tab:1}}
>
> Looks like we failed to compile the PDF

@arfon @cMadan
The paper successfully compiled on my machine using Pandoc. However, whedon cannot compile the paper and produced above error.
