### 2019-01-21 56987cb
Added donations to the README file [skip ci].
