### 2019-01-24 #1155 @raamana
Thanks Chris, happy to review, this is an interesting and useful tool. However, due to prior commitments etc, I may take 4-8 weeks.
