### 2019-01-24 7ef13ea
Added a donate button. [skip ci]
