### 2019-02-01 6535bab
Added rectangular feature to OVO-TSVM estimator.
