### 2019-02-14 #1155 @labarba
I wonder if @stefanv could help here ... Stefan: are you able to review this submission for JOSS?
