### 2019-02-14 #1155 @stefanv
- [x] **Repository:** Is the source code for this software available at the repository url?

Yes, but I wanted to verify whether this license also applies to the included optimizer.

Potential license breach through publication of datasets without attribution:

https://archive.ics.uci.edu/ml/citation_policy.html

- [ ] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.13750">software archive</a> resolve?

No.  The installation process is non-standard for a Python library.  On my (Debian) system, the shell script (which I was hesitant to run, as I imagine many users will be), did not correctly detect Python 3 and and then asked for sudo access (at which point I aborted).

Please provide clear steps for installing manually from source.

I will proceed with these items as soon as I am able to install.

- [ ] **Performance:** Have the performance claims of the software been confirmed?

- [ ] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

I could not locate the API documentation.

Yes, but will verify once installation is completed.

None that I could find.

Editor: is this a requirement of the `.bib` file?  Not entirely sure what is meant by "archival reference".  No DOIs are provided, as far as I can tell.
