### 2019-02-15 #1155 @mir-am
Hi

We thank the reviewer for his valuable comments and suggestions.
We will carefully revise the paper and LightTwinSVM program according to the reviewer's comments.

We share our responses to the comments as soon as all the reviewer's comments are addressed.

Regards,
Mir
