### 2019-02-15 234ed15
The JOSS paper is revised according the reviewer's comments. [skip ci]
