### 2019-02-15 #1155 @labarba
Folks, this is the Pre-Review issue. We need to move to the Review issue for these conversations. I didn't hear back from @rougier, so let's rely on the expert review by @stefanv, and let me start the Review now.
