### 2019-02-15 #1252 @labarba
@stefanv — Can I ask you to copy over your review comments here?
https://github.com/openjournals/joss-reviews/issues/1155#issuecomment-463760234
