### 2019-02-15 #1252 @labarba
This one too
https://github.com/openjournals/joss-reviews/issues/1155#issuecomment-463750224
