### 2019-02-15 #1252 @labarba
_Why_, you may ask? We're looking into ways of exporting the Review issues into an archival form, to preserve and link to the published articles. (But not the Pre-Review issue, whose purpose is to find editors and reviewers, or occasionally make a desk rejection if the submission is out of scope).

Also, @stefan, you have a Reviewer checklist here, and you have automatically been added as a collaborator to the repo, so you can check items off there.
