### 2019-02-15 #1252 @labarba
On this question, @stefanv

> What is JOSS policy on content recommendations? E.g., I have some recommended language edits, as well as experiment suggestion (since there is a benchmark included in the paper).

You should feel free to make content recommendations, and also language edits (I do these all the time!).
