### 2019-02-15 #1252 @labarba
Your checklist is now added. If you have reviewed before for JOSS, you should already have repo permissions to edit the checklist. Let me know if you run into any trouble.
