### 2019-02-15 #1155 @labarba
Yes, you are right. But I needed an affirmation from you to be added as reviewer. If I add you in the Pre-Review issue, then `whedon` automatically adds a review checklist for you in the Review issue, and adds permissions for you to edit (if you don't have them already).
