### 2019-02-15 mir-am/LightTwinSVM#3 @rougier
Installation problem on OSX
I tried to run the `setup.sh` on OSX (10.14.2) using `Python 3.7.2`  but the script failed without much information. Is there a way to have more debug information?

```
The Installation of LightTwinSVM program:
Make sure that you are connected to the internet so that dependencies can be downloaded.
Press enter to start the installation process...
***************************************
Step 1:
Python 3.5 or newer detected on your system...
pip tool is detected...
***************************************
Step 2:
Installs required Python packages to run the program...
Requirement already satisfied: numpy in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 2)) (1.15.2)
Requirement already satisfied: scipy in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 3)) (1.1.0)
Requirement already satisfied: scikit-learn in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 4)) (0.19.2)
Requirement already satisfied: pandas in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 5)) (0.23.4)
Collecting pybind11 (from -r requirments.txt (line 6))
  Downloading https://files.pythonhosted.org/packages/f2/7c/e71995e59e108799800cb0fce6c4b4927914d7eada0723dd20bae3b51786/pybind11-2.2.4-py2.py3-none-any.whl (145kB)
    100% |████████████████████████████████| 153kB 4.8MB/s
Collecting xlsxwriter (from -r requirments.txt (line 7))
  Downloading https://files.pythonhosted.org/packages/8d/d0/f4ee9c621902d18807dd88b0a0ff4426666af8dd1e952c4943a9415f3eeb/XlsxWriter-1.1.4-py2.py3-none-any.whl (142kB)
    100% |████████████████████████████████| 143kB 5.3MB/s
Requirement already satisfied: sklearn in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 9)) (0.0)
Requirement already satisfied: pytz>=2011k in /usr/local/lib/python3.7/site-packages (from pandas->-r requirments.txt (line 5)) (2018.5)
Requirement already satisfied: python-dateutil>=2.5.0 in /usr/local/lib/python3.7/site-packages (from pandas->-r requirments.txt (line 5)) (2.7.3)
Requirement already satisfied: six>=1.5 in /usr/local/lib/python3.7/site-packages (from python-dateutil>=2.5.0->pandas->-r requirments.txt (line 5)) (1.11.0)
Installing collected packages: pybind11, xlsxwriter
Successfully installed pybind11-2.2.4 xlsxwriter-1.1.4
Checks existence of Python's Tkinter...
Found python3-tk on your system...
Looking for Python 3 dev. package...
Found Python 3 dev on your system...
***************************************

Step 3:
Checks existence of LAPACK and BLAS...
Cloning into 'temp'...
remote: Enumerating objects: 1, done.
remote: Total 4513 (delta 0), reused 0 (delta 0), pack-reused 4512
Resolving deltas: 100% (2920/2920), done.
The C++ extension moudle is generated...
***************************************

The installation failed... Number of compeleted steps: 6
```
