### 2019-02-15 #1252 @stefanv
- **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

Yes, but I wanted to verify whether this license also applies to the included optimizer.

Potential license breach through publication of datasets without attribution:

https://archive.ics.uci.edu/ml/citation_policy.html

- **Installation:** Does installation proceed as outlined in the documentation?

No.  The installation process is non-standard for a Python library.  On my (Debian) system, the shell script (which I was hesitant to run, as I imagine many users will be), did not correctly detect Python 3 and and then asked for sudo access (at which point I aborted).

Please provide clear steps for installing manually from source.

- **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

I could not locate the API documentation.

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

None that I could find.

- **References:** Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

Editor: is this a requirement of the `.bib` file?  Not entirely sure what is meant by "archival reference".  No DOIs are provided, as far as I can tell.
