### 2019-02-15 #1252 @rougier
I've started the review (and post an issue on the repo because I cannot build it). In the meantime and to parallelize things here are a few comments and questions:

* In the paper, the legend of the figures 1 & 2 could benefit from a better caption. As they stand, they do not really help i finding what is to be seen.
* who is the second author (since you have made all the commits, what is his role in this software?)
* for all your references, you need to add the DOI
* The README is great, but where is the API documentation, did you upload a documentation to readthedocs ? (your doc Makefile seems to indicate that you can build the doc, it would be helpful ot be able to read it online)
