### 2019-02-15 #1252 @cMadan
@stefanv @rougier, thank you both for agreeing to review this submission! Where relevant, please direct your JOSS-related questions to me, as the editor for this submission. @labarba helped out here in getting things started, but she's one of the Associated Editor-in-Chief's and edits other submissions, so we don't want to have her answer all the questions--that's part of my role here.

@labarba, thanks so much for getting things started and sorting out the review-related comments that were in the pre-review. I'll try and take it from here :).

@stefanv, DOIs are required in the bib (when possible), so, yes, those need to be added.

@rougier, that's a good suggestion, thanks! We can see about updating the initial review comment to include that.
