### 2019-02-17 715df9e
Added more debug information to setup.sh file for fixing the installation issue on OSX systems.
