### 2019-02-17 mir-am/LightTwinSVM#3 @mir-am
Hi

Thank you for reviewing the LightTwinSVM program.
Based on the above output, the installation is almost completed. Looks like the last step failed which is compiling the C++ extension module (optimizer).
My first guess is that the setup script fails to detect your OSX system. Therefore, the C++ extension cannot be generated.
I think $OSTYPE variable is different on your OSX system from what is hard-coded in the setup script.
Anyhow, I have added more debug information to setup.sh file so that we can fix the installation issue on your system.
Please run the setup.sh file again and copy-paste the output of setup script here.

Regards,
Mir
