### 2019-02-17 mir-am/LightTwinSVM#3 @rougier
I've update my copy, but still the same cryptic final message. Is there an option for setup debug?

```
The Installation of LightTwinSVM program:
Make sure that you are connected to the internet so that dependencies can be downloaded.
Press enter to start the installation process...
***************************************
Step 1:
Python 3.5 or newer detected on your system...
Python version: Python 3.7.2
OS: darwin18
pip tool is detected...
***************************************
Step 2:
Installs required Python packages to run the program...
Requirement already satisfied: numpy in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 2)) (1.15.2)
Requirement already satisfied: scipy in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 3)) (1.1.0)
Requirement already satisfied: scikit-learn in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 4)) (0.19.2)
Requirement already satisfied: pandas in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 5)) (0.23.4)
Requirement already satisfied: pybind11 in /Users/rougier/Library/Python/3.7/lib/python/site-packages (from -r requirments.txt (line 6)) (2.2.4)
Requirement already satisfied: xlsxwriter in /Users/rougier/Library/Python/3.7/lib/python/site-packages (from -r requirments.txt (line 7)) (1.1.4)
Requirement already satisfied: sklearn in /usr/local/lib/python3.7/site-packages (from -r requirments.txt (line 9)) (0.0)
Requirement already satisfied: pytz>=2011k in /usr/local/lib/python3.7/site-packages (from pandas->-r requirments.txt (line 5)) (2018.5)
Requirement already satisfied: python-dateutil>=2.5.0 in /usr/local/lib/python3.7/site-packages (from pandas->-r requirments.txt (line 5)) (2.7.3)
Requirement already satisfied: six>=1.5 in /usr/local/lib/python3.7/site-packages (from python-dateutil>=2.5.0->pandas->-r requirments.txt (line 5)) (1.11.0)
Checks existence of Python's Tkinter...
Found python3-tk on your system...
Looking for Python 3 dev. package...
Found Python 3 dev on your system...
***************************************

Step 3:
Checks existence of LAPACK and BLAS...
Found Armadillo repository. No need to clone again.
Failed to generate the C++ extension module...
***************************************

The installation failed... Number of steps completed: 5
```
