### 2019-02-19 #1252 @rougier
Quick update on the review, we're trying to solve an [installation problem](https://github.com/mir-am/LightTwinSVM/issues/3)
