### 2019-02-19 c07554b
DOIs are added to the bib file of the JOSS paper. [skip ci]
