+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={Understanding machine learning: From theory to algorithms},
   author={Shalev-Shwartz, Shai and Ben-David, Shai},
   year={2014},
-  publisher={Cambridge university press}
+  publisher={Cambridge university press},
+  doi={10.1017/CBO9781107298019}
 }
 
 @article{kotsiantis2007,
@@ -22,7 +23,8 @@
    number = {5},
    ISSN = {0162-8828},
    year = {2007},
-   type = {Journal Article}
+   type = {Journal Article},
+   doi={10.1109/TPAMI.2007.1068}
 }
 
 @article{chang2011libsvm,
@@ -33,7 +35,8 @@
   number={3},
   pages={27},
   year={2011},
-  publisher={Acm}
+  publisher={Acm},
+  doi={10.1145/1961189.1961199}
 }
 
 @article{pedregosa2011scikit,
@@ -54,7 +57,8 @@
    pages = {266-278},
    ISSN = {0950-7051},
    year = {2014},
-   type = {Journal Article}
+   type = {Journal Article},
+   doi={10.1016/j.knosys.2014.08.005}
 }
 
 @article{mir2018,
@@ -69,6 +73,7 @@
    number="12",
    pages="4551--4564",
    issn="1573-7497",
+   doi="10.1007/s10489-018-1225-z"
 }
 
 @article{walt2011,
@@ -80,7 +85,8 @@
    pages = {22-30},
    ISSN = {1521-9615},
    year = {2011},
-   type = {Journal Article}
+   type = {Journal Article},
+   doi={10.1109/MCSE.2011.37}
 }
 
 @article{mckinney2011pandas,
