### 2019-02-22 e820dde
README file updated with a section for building the program manually. [skip ci]
