### 2019-02-22 #1252 @mir-am
Hi

To help the reviewers and also users install and build the program manually, a step-by-step guide is provided [here](https://github.com/mir-am/LightTwinSVM/wiki/Building-the-LightTwinSVM-manually-on-Linux-and-OSX-systems) in the project's wiki.
Moreover, DOIs are added to all the references in the bib file that have one.

We will also address other comments of the reviewers as soon as possible.

Regards,
Mir
