### 2019-02-22 mir-am/LightTwinSVM#3 @mir-am
For now, a step-by-step guide is provided [here](https://github.com/mir-am/LightTwinSVM/wiki/Building-the-LightTwinSVM-manually-on-Linux-and-OSX-systems) to build the program manually.

Also, I try to fix the setup script for OSX systems. Based on the above output, there is a problem in conditions of if statements in the bash script.

https://github.com/mir-am/LightTwinSVM/blob/e820dde718c9675bcf7279e0bf537cc2acfc76f4/setup.sh#L183

Of course, a simple solution would be creating a separate shell script for OSX systems. However, there are some shared installation steps for both OSX and Linux systems, which I think it is better to have one setup script for both systems.
