### 2019-02-22 #1252 @stefanv
@mir-am Thank you for these instructions; that's helpful already.  Do you have plans to build binary packages?  If so, you may want to look at https://github.com/matthew-brett/multibuild. This would make your package pip-installable.  That, and setting up a recipe on conda-forge, will give you access to the majority of Python users out there.
