+++ docs/source/index.rst
@@ -10,7 +10,7 @@ Welcome to LightTwinSVM's documentation!
    :maxdepth: 2
    :caption: Contents:
 
-
+   modules
 
 Indices and tables
 ==================
