+++ docs/source/modules.rst
@@ -1,0 +1,12 @@
+Modules of LightTwinSVM's package
+===============
+
+This page contains the list of the project's modules
+
+.. autosummary::
+   :toctree: _autosummary
+
+   twinsvm
+   eval_classifier
+   dataproc
+   misc
