### 2019-02-24 599d34c
Sphinx conf file updated. Also, comments are changed in source files. [skip ci]
