+++ ltsvm/twinsvm.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
 
-"""
-LightTwinSVM Program - Simple and Fast
-Version: 0.2.0-alpha - 2018-05-30
-Developer: Mir, A. (mir-am@hotmail.com)
-License: GNU General Public License v3.0
+# LightTwinSVM Program - Simple and Fast
+# Version: 0.5.0 - 2019-01-01
+# Developer: Mir, A. (mir-am@hotmail.com)
+# License: GNU General Public License v3.0
 
-Module: twinsvm.py
-In this module, functios is defined for training and testing TwinSVM classifier.
+"""
+Classes and functios are defined for training and testing TwinSVM classifier.
 
 TwinSVM classifier generates two non-parallel hyperplanes.
 For more info, refer to the original papar.
