+++ ltsvm/dataproc.py
@@ -1,15 +1,12 @@
 # -*- coding: utf-8 -*-
 
+# LightTwinSVM Program - Simple and Fast
+# Version: 0.5.0 - 2019-01-01
+# Developer: Mir, A. (mir-am@hotmail.com)
+# License: GNU General Public License v3.0
 
 """
-LightTwinSVM Program - Simple and Fast
-Version: 0.2.0-alpha - 2018-05-30
-Developer: Mir, A. (mir-am@hotmail.com)
-License: GNU General Public License v3.0
-
-Module: dataproc.py
 In this module, functions for reading and processing datasets are defined.
-
 """
 
 
