### 2019-02-26 mir-am/LightTwinSVM#3 @rougier
Now I get
```
 $  g++ -O3 -Wall -shared -std=c++11 -fPIC `python3 -m pybind11 --includes` ./ltsvm/optimizer/pybind_clippdcd.cpp -o ./ltsvm/optimizer/clipdcd`python3-config --extension-suffix` -I ./temp/include -DARMA_DONT_USE_WRAPPER -framework Accelerate
/anaconda3/bin/python3: No module named pybind11
./ltsvm/optimizer/pybind_clippdcd.cpp:27:10: fatal error: 'pybind11/pybind11.h' file not found
#include <pybind11/pybind11.h>
         ^~~~~~~~~~~~~~~~~~~~~
1 error generated.
```
