### 2019-02-26 mir-am/LightTwinSVM#3 @rougier
After having installed pybind11 via Homebrew, compilation still doesn't work because of a missing Python.h (that is installed). I suspect the unusual command line might be the cause
