### 2019-02-26 #1252 @rougier
And I'm still fighting with the installation procedure (with no luck yet)
