### 2019-02-26 97590a7
Fixed compliation and linking error of C++ extension module for OSX. [skip ci]
