### 2019-02-26 mir-am/LightTwinSVM#3 @mir-am
> After having installed pybind11 (via Homebrew), compilation still doesn't work with a lot of undefined symbols that seem to indicate some base libraries (Python) are not compatible. Do you have an alternative way to install it?

pybind11 is included in the "requirements.txt" file, which will be installed by pip. Anyhow, I fixed compilation error for OSX. A different compiler flag should have added to g++ command for OSX, which solves the linking error (undefined symbols).
Please compile the C++ extension using the following command:

```
g++ -O3 -Wall -shared -std=c++11 -undefined dynamic_lookup `python3 -m pybind11 --includes` ./ltsvm/optimizer/pybind_clippdcd.cpp -o ./ltsvm/optimizer/clipdcd`python3-config --extension-suffix` -I ./temp/include -DARMA_DONT_USE_WRAPPER -framework Accelerate
```

I have tested the above command and it successfully compiles the extenstion module on OSX machine of Travis CI.

If you received another error during installation, please let me know.
