### 2019-02-26 mir-am/LightTwinSVM#3 @rougier
I managed to compile using your line but the setup.sh is still broken.
