### 2019-02-27 bb896dd
Fixed build issue for OSX on TravisCI.
