+++ .travis.yml
@@ -23,10 +23,9 @@ matrix:
       sudo: true
       python: '3.7'
       
-  allow_failures:
     - os: osx
       language: generic
-      env: PYENV_VERSION=3.6.7
+      env: PYENV_VERSION=3.7.2
 
 branches:
   only:
@@ -44,10 +43,9 @@ addons:
 before_install:
   - chmod +x install-travisci.sh
   - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then echo $'#!/bin/bash\npython3 -m unittest discover -s tests' > ltsvm_test.sh; fi
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then echo $'#!/bin/bash\npython -m unittest discover -s tests' > ltsvm_test.sh; fi
   - chmod +x ltsvm_test.sh
   
 install: ./install-travisci.sh
 
 script:
-- ./ltsvm_test.sh
+- if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then ./ltsvm_test.sh; fi
