### 2019-02-27 0afe52f
Fixed travis.yml file that caused OSX build to fail. [skip appveyor]
