### 2019-02-28 99ea840
Fixed building API docs on the Read the Docs.
