### 2019-03-01 72838b3
API docs of twinsvm.py module revised according to the NumPydoc.
