+++ ltsvm/twinsvm.py
@@ -28,16 +28,51 @@ from ltsvm.optimizer import clipdcd
 
 
 class TSVM(BaseEstimator):
+    
+    """
+    Twin Support Vector Machine for binary classification.
+    
+    Parameters
+    ----------
+    kernel : str, optional (default='linear')
+        Type of the kernel function which is either 'linear' or 'RBF'.
+    
+    rect_kernel : float, optional (default=1.0)
+        Percentage of training samples for Rectangular kernel.
+        
+    C1 : float, optional (default=1.0)
+        Penalty parameter of first optimization problem.
+        
+    C2 : float, optional (default=1.0)
+        Penalty parameter of second optimization problem.
+        
+    gamma : float, optional (default=1.0)
+        Parameter of the RBF kernel function.
+    
+    Attributes
+    ----------
+    mat_C_t : array-like, shape = [n_samples, n_samples]
+        A matrix that contains kernel values.
+        
+    cls_name : str
+        Name of the classifier.
+    
+    w1 : array-like, shape=[n_features]
+        Weight vector of class +1's hyperplane.
+        
+    b1 : float
+        Bias of class +1's hyperplane.
+        
+    w2 : array-like, shape=[n_features]
+        Weight vector of class -1's hyperplane.
+    
+    b2 : float
+        Bias of class -1's hyperplane.
+    
+    """
 
     def __init__(self, kernel='linear', rect_kernel=1, C1=2**0, C2=2**0, \
                  gamma=2**0):
-
-        """
-        Input:
-            Kernel_type: 1->Linear, 2->RBF(Gaussion)
-            c1, c2: Penalty parameters
-            gamma: RBF function parameter
-        """
 
         self.C1 = C1
         self.C2 = C2
@@ -53,7 +88,13 @@ class TSVM(BaseEstimator):
     def get_params_names(self):
         
         """
-        It returns the names of hyper-parameters of this classifier.
+        For retrieving the names of hyper-parameters of this classifier.
+        
+        Returns
+        -------
+        parameters : list of str, {['C1', 'C2'], ['C1', 'C2', 'gamma']}
+            Returns the names of the hyperparameters which are same as
+            the class' attributes.
         """
         
         return ['C1', 'C2'] if self.kernel == 'linear' else ['C1', 'C2', 'gamma']
@@ -61,15 +102,18 @@ class TSVM(BaseEstimator):
     def fit(self, X_train, y_train):
 
         """
-        It trains TwinSVM classfier on given data
-        Input:    
-            X_train: Training samples
-            y_train: Samples' category
-        output:
-
-            w1, w2: Coordinates of two non-parallel hyperplanes
-            b1, b2: Biases
-            """
+        It fits the binary TwinSVM model according to the given training data.
+        
+        Parameters
+        ----------
+        X_train : array-like, shape (n_samples, n_features) 
+           Training feature vectors, where n_samples is the number of samples
+           and n_features is the number of features. 
+           
+        y_train : array-like, shape(n_samples,)
+            Target values or class labels.
+           
+        """
 
         # Matrix A or class 1 samples
         mat_A = X_train[y_train == 1]
@@ -134,9 +178,18 @@ class TSVM(BaseEstimator):
     def predict(self, X_test):
 
         """
-            Predictes class of test samples
-            Input:
-                X_test: Test samples    
+        Performs classification on samples in X using the TwinSVM model.
+        
+        Parameters
+        ----------
+        X_test : array-like, shape (n_samples, n_features)
+            Feature vectors of test data.
+                
+        Returns
+        -------
+        output : array, shape (n_samples,)
+            Predicted class lables of test data.
+            
         """
 
[…]
