### 2019-03-01 5c3c284
Added numpydoc to requirements file. Also, docs badge added to README file. [skip ci]
