+++ requirments.txt
@@ -1,5 +1,6 @@
 # The folliwing Python pacakges is needed for building and running LightTwinSVM program
 numpy
+numpydoc
 scipy
 scikit-learn
 pandas
