### 2019-03-01 4ff76c9
Installing NumPydoc ver. 0.7.0 in requirements.txt file for properly displaying docs. [skip ci]
