+++ requirments.txt
@@ -1,6 +1,7 @@
 # The folliwing Python pacakges is needed for building and running LightTwinSVM program
 numpy
-numpydoc
+# NumPydoc ver. 0.7.0 should be install for building docs properly
+numpydoc==0.7.0
 scipy
 scikit-learn
 pandas
