### 2019-03-02 ab02237
Revised API docs of dataproc.py module using NumPydoc. [skip ci]
