### 2019-03-02 5fa1c26
Revised API docs of eval_classifier.py module according to the NumPydoc. Part 1. [skip ci].
