### 2019-03-03 79d80ac
Revised API docs of eval_classifier.py module according to the NumPydoc-Final part. [skip ci]
