+++ ltsvm/eval_classifier.py
@@ -7,8 +7,8 @@
 # License: GNU General Public License v3.0
 
 """
-In this module, classes and methods are defined for evluating the performance of the
-TwinSVM model. Also, a method for saving detailed classification result.
+In this module, classes and methods are defined for evluating the performance
+of the TwinSVM model. Also, a method for saving detailed classification result.
 """
 
 from ltsvm.twinsvm import TSVM, MCTSVM, OVO_TSVM
@@ -26,8 +26,8 @@ import os
 def eval_metrics(y_true, y_pred):
 
     """
-    It evaluates the performance of the TwinSVM model based on common evaluation
-    metrics such Accuracy, Recall, Precision, and F1-measure.
+    It computes common evaluation metrics such as Accuracy, Recall, Precision,
+    F1-measure, and elements of the confusion matrix.
     
     Parameters
     ----------
@@ -138,7 +138,7 @@ def eval_metrics(y_true, y_pred):
 class Validator:
 
     """
-    It evaluates the TwinSVM model based on a evaluation method.
+    It evaluates the TwinSVM model based on the specified evaluation method.
     
     Parameters
     ----------
@@ -155,8 +155,8 @@ class Validator:
         ('t_t_split', 30) -> 30% of samples for test set.
         
     obj_tsvm : object
-        A TwinSVM model. It can be an instace of :class:`TSVM`, :class:`MCTSVM`,
-        or :class:`OVO_TSVM`.
+        A TwinSVM model. It can be an instace of :class:`TSVM <twinsvm.TSVM>`,
+        :class:`MCTSVM <twinsvm.TSVM>`, or :class:`OVO_TSVM <twinsvm.TSVM>`.
     """
 
     def __init__(self, X_train, y_train, validator_type, obj_tsvm):
@@ -169,7 +169,7 @@ class Validator:
     def cv_validator(self, dict_param):
 
         """
-        It evaluates the TwinSVM model using the cross-validaion method.
+        It evaluates the TwinSVM model using the cross-validation method.
         
         Parameters
         ----------
@@ -262,7 +262,7 @@ class Validator:
         Returns
         -------
         float
-            accuracy of the model.
+            Accuracy of the model.
             
         float
             Zero standard deviation.
@@ -297,7 +297,23 @@ class Validator:
     def cv_validator_mc(self, dict_param):
 
         """
-        It applies cross validation on instance of multiclass TSVM classifier
+        It evaluates the multi-class TwinSVM model using the cross-validation.
+        
+        Parameters
+        ----------
+        dict_param : dict 
+            Values of hyper-parameters for the multiclss TwinSVM model.
+            
+        Returns
+        -------
+        float
+            Accuracy of the model.
+            
+        float
+            Zero standard deviation.
+            
+        dict
+            Evaluation metrics such as Recall, Percision and F1-measure.
         """
 
         # Set parameters of multiclass TSVM classifer
@@ -343,7 +359,13 @@ class Validator:
     def choose_validator(self):
 
         """
-        It returns choosen validator method.
+        It selects the appropriate evaluation method based on the input
+        paramters.
+        
+        Returns
+        -------
+        object
+            An evaluation method for assesing the model's performance.
         """
 
         if isinstance(self.obj_TSVM, TSVM):  # Binary TSVM
@@ -367,13 +389,44 @@ def search_space(kernel_type, class_type, c_l_bound, c_u_bound, rbf_lbound, \
                  rbf_ubound, step=1):
 
     """
-    It generates combination of search elements for grid search
-    Input:
-        kernel_type: kernel function which is either linear or RBF
-        c_l_bound, c_u_bound: Range of C penalty parameter for grid search(e.g 2^-5 to 2^+5)
[…]
