### 2019-03-03 068be46
Revised API docs of misc.py module according to the NumPydoc. [skip ci]
