### 2019-03-03 18febfe
README file updated with API documentation.
