+++ docs/source/index.rst
@@ -5,6 +5,7 @@
 
 Welcome to LightTwinSVM's documentation!
 ========================================
+You can read API documentation, if you want to use the LightTwinSVM's estimators and tools directly in your project. Furthermore, if you have a question or problem regarding this program, please see the `support section <https://github.com/mir-am/LightTwinSVM#support>`_ of the project in its GitHub repo.
 
 .. toctree::
    :maxdepth: 2
