### 2019-03-04 b13e1dd
A contributing guide for the LightTwinSVM project. [skip ci]
