### 2019-03-05 fd397bc
wrote documentation section for CONTRIBUTUNG guide. [skip ci]
