### 2019-03-06 bfb202f
Corrected grammar for contributing guide. [skip ci]
