### 2019-03-06 dd2c930
acknowledged the use of UCI datasets. [skip ci]
