### 2019-03-06 26bf67e
Fixed a typo error. Also new info added to contributing guide for documentation. [skip ci]
