+++ CONTRIBUTING.md
@@ -30,4 +30,7 @@ To develop a new feature or fix a bug, please do the following steps:
 - By having an approved feature or a bug fix, you will become one of the contributors of the LightTwinSVM project.
 
 ## Documentation
-To contribute through documentation, you can fix typo errors, or you may want to improve API docs by adding usage examples. For instance, you can create a new page in the [documentation](https://lighttwinsvm.readthedocs.io/en/latest/), which describes how to use the LightTwinSVM API to solve a classification task. As of writing this, the project's documentation lacks good usage examples which shows power users how to use TwinSVM estimators for their project/problem. Nonetheless, if you have a new idea on the improvement of the docs, let me know by visiting the [support section](https://github.com/mir-am/LightTwinSVM#support) for contact details.
+To contribute through documentation, you can fix typo errors, or you may want to improve API docs by adding usage examples. For instance, you can create a new page in the [documentation](https://lighttwinsvm.readthedocs.io/en/latest/), which describes how to use the LightTwinSVM API to solve a classification task. As of writing this, the project's documentation lacks good usage examples which shows power users how to use TwinSVM estimators for their project/problem. Nonetheless, if you have a new idea on the improvement of the docs, let me know. (See the [support section](https://github.com/mir-am/LightTwinSVM#support) for contact details.)
+
+### Guide
+To improve the documentation, make your changes in the [docs](https://github.com/mir-am/LightTwinSVM/tree/master/docs) folder. Note that this project uses [Sphinx](http://sphinx-doc.org/) and its API doc follows the [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html) rules. After you made your changes, commit them, and then create [a pull request].
