### 2019-03-07 #1252 @mir-am
Also, we would like to thank @rougier for his comments and reviewing this work.

Our responses to the comments are as follows:

>In the paper, the legend of the figures 1 & 2 could benefit from a better caption. As they stand, they do not really help i finding what is to be seen.

In the research field of SVM, expressions like "Geometric illustration" or "Geometric interpretation" are common. These captions are used to show the main idea of a classifier in 2-dimensional feature space. Nonetheless, we welcome suggestions for the captions' title.

> who is the second author (since you have made all the commits, what is his role in this software?)

He is my supervisor and taught me the theories behind SVM and its extensions like TwinSVM. Moreover, the LightTwinSVM project is developed at his lab. Therefore, I must include his name in the paper.

> for all your references, you need to add the DOI

DOIs were added to references in the paper when possible.
https://github.com/mir-am/LightTwinSVM/commit/c07554bf0f98a75f76025d6d3d5e8f3946615623

> The README is great, but where is the API documentation, did you upload a documentation to readthedocs ? (your doc Makefile seems to indicate that you can build the doc, it would be helpful to be able to read it online)

Yes. The LightTwinSVM's documentation was uploaded to Read the Docs. The API documentation can be read [here](https://lighttwinsvm.readthedocs.io/en/latest/index.html). All the LightTwinSVM's estimators and tools are documented so that users can employ them for their projects.
