### 2019-03-08 #1252 @rougier
Thanks. I'm still fighting with the installation on OSX bur we've made some progress.
