### 2019-03-09 mir-am/LightTwinSVM#3 @mir-am
> Any idea ?

Unfortunately, I don't have access to Mac OSX system to test the setup script and then fix it. However, I might be able to solve the issue using Travis CI's OSX machines. Therefore, it may take several days.

To speed up the review process, please follow the [step-by-step guide](https://github.com/mir-am/LightTwinSVM/wiki/Building-the-LightTwinSVM-manually-on-Linux-and-OSX-systems) to test and review the LightTwinSVM program.

After all, I will fix this issue as soon as possible.
