### 2019-03-09 mir-am/LightTwinSVM#3 @rougier
The last release seems to be working, at least for the manual instruction. `setup.sh` still fails but I imagine it is now easier to fix. Test also passed. I'm closing the issue.
