### 2019-03-09 #1252 @rougier
Some update on the installation problem: it has been fixed. @cMadan can you rebuild the PDF?
