### 2019-03-09 b7da866
Fixed setup.sh issue on OSX. [skip ci]
