### 2019-03-09 mir-am/LightTwinSVM#3 @mir-am
@rougier
I managed to fix the setup.sh issue on Travis CI's virtual OSX system. The issue was somehow a hidden syntax problem!
I appreciate if you clone the repo again and run the setup.sh script. Thanks.
