### 2019-03-10 mir-am/LightTwinSVM#3 @rougier
Just did it and it works now.
