### 2019-03-11 #1252 @rougier
Thanks. I'm almost done with the review. @mir-am There are still some missing DOI in the references, can you check? For the captions of figure 1 & 2,  maybe you can have figures side by side in one figure and explain in the caption the fundamental difference between SVM and TwinSVM.
