### 2019-03-12 #1252 @mir-am
> Thanks. I'm almost done with the review. @mir-am There are still some missing DOI in the references, can you check? For the captions of figure 1 & 2, maybe you can have figures side by side in one figure and explain in the caption the fundamental difference between SVM and TwinSVM (but if your prefer to leave them as they are, I'm fine).

Thank you for reviewing this work.
I tried again to find missing DOI for some references. Looks like The Journal of Machine Learning research does not assign a DOI to its papers.
Your suggestion about the figures seems fine for me. I think the paper's width is not large enough to place figures side by side and it may make the comparison fruitless.  However, I couldn't use subfigures feature to place figures below each other.

@cMadan
How can I place figures below each other like subfigures in LaTeX? I appreciate your help.
