### 2019-03-13 mir-am/LightTwinSVM#4 @stefanv
`requirements.txt` misnamed
The file is currently called `requirments.txt`.

xref: https://github.com/openjournals/joss-reviews/issues/1252
