### 2019-03-13 #1252 @stefanv
> We will run the benchmark again and report the standard deviation of the accuracy in the paper.

Apart from this one outstanding item, I think this is good to go.
