### 2019-03-14 #1252 @cMadan
@mir-am, unfortunately, you can't do subfigures in the paper markdown. Instead, you should combine the figures using a graphics program and add the "(a)" panel notation, such that the figures are a single image.
