### 2019-03-17 a24e1e1
Corrected typo and grammar erros in the JOSS paper [skip ci]
