### 2019-03-17 mir-am/LightTwinSVM#4 @mir-am
Thank you for creating this issue.
Actually, I was aware of it!

I am going to correct the name of the file now.
