+++ .appveyor.yml
@@ -25,7 +25,7 @@ install:
   - "pip --version"
   
   # Install dependencies for testing LightTwinSVM
-  - "pip install -r requirments.txt"
+  - "pip install -r requirements.txt"
   
   # Install Cython for building extension module
   - "pip install Cython"
