### 2019-03-17 #1252 @mir-am
> @mir-am, unfortunately, you can't do subfigures in the paper markdown. Instead, you should combine the figures using a graphics program and add the "(a)" panel notation, such that the figures are a single image.

Thank you for help and suggestion.
The two figures are on the same page. Therefore, readers can see the difference between the two classifiers. However, we wanted to use subfigures in order to combine the two figures into one.
Anyhow, in my opinion, the location of the two figures on the page is fine.

This issue was also fixed without breaking the installation scripts for both CI services and users. (Thanks to @stefanv for reminding us of the issue.)
https://github.com/mir-am/LightTwinSVM/issues/4
