### 2019-03-17 mir-am/LightTwinSVM#4 @mir-am
The issue was fixed without breaking the installation scripts.
I may now close the issue.
