### 2019-03-23 #1252 @cMadan
@mir-am, do you plan to report the SD of accuracy in the paper, as @stefanv recommended?

@stefanv, do you have any other remaining concerns about the performance of the package? (It looks like this is the only thing in the reviewer checklist you haven't checked off yet.)
