### 2019-03-23 #1252 @mir-am
> @mir-am, do you plan to report the SD of accuracy in the paper, as @stefanv recommended?
>

Yes. As of writing this, I am conducting experiments to report the SD of accuracies.
I will report the results in the paper as soon as all the experiments are finished.
I will have it done within a few days.
