### 2019-03-25 bf0a4f0
Revised the JOSS paper to include the SD of accuracies. [skip ci]
