### 2019-03-25 #1252 @mir-am
@cMadan @stefanv
We carefully revised the paper for including the new experiments (as requested by the reviewer).

Modifications to the paper are as follows:

- A new section “Numerical experiments” was created to analyze the performance of LightTwinSVM and scikit-learn’s SVM.
- To add the standard deviation to all accuracies in Table 1, we conducted experiments again for both classifiers. However, new results are reported for some datasets. The accuracy for some datasets slightly improved for LightTwinSVM. This accuracy improvement is due to the choosing optimal values of hyper-parameters from a specified range, which is stated in the paper.
- We also wrote more detail about hyper-parameter selection in the section of “numerical experiments”.

We welcome new suggestions and ideas to improve the LightTwinSVM and its paper.
