### 2019-03-29 #1252 @cMadan
@mir-am, thinks look good and I'm almost ready to accept. Can you mint a new release that includes the changes from the review process? I also need a DOI for an archived version of the code (i.e., from Zenodo or figshare).
