### 2019-03-31 0013126
Updated the results of Numerical Experiments in README file. To include the results of the JOSS paper. [skip ci]
