### 2019-03-31 ee3c9ab
CHANGELOG file updated with new version of the program. [skip ci]
