+++ ltsvm/__init__.py
@@ -2,13 +2,13 @@
 # -*- coding: utf-8 -*-
 """
 LightTwinSVM Program - Simple and Fast
-Version: 0.3.0-alpha - 2018-06-03
+Version: 0.6.0 - 2019-03-31
 Developer: Mir, A. (mir-am@hotmail.com)
 License: GNU General Public License v3.0
 
 A place for package constants
 """
 
-__version__ = '0.5.0'
-RELEASE_DATE = '2019-01-01'
+__version__ = '0.6.0'
+RELEASE_DATE = '2019-03-31'
 
