+++ docs/source/conf.py
@@ -38,9 +38,9 @@ copyright = '2019, Mir, A.'
 author = 'Mir, A.'
 
 # The short X.Y version
-version = '0.5.0'
+version = '0.6.0'
 # The full version, including alpha/beta/rc tags
-release = '0.5.0'
+release = '0.6.0'
 
 
 # -- General configuration ---------------------------------------------------
