### 2019-03-31 282ccbd
Fixed typo errors in CHANGELOG file.
