### 2019-03-31 #1252 @mir-am
> @mir-am, thinks look good and I'm almost ready to accept. Can you mint a new release that includes the changes from the review process? I also need a DOI for an archived version of the code (i.e., from Zenodo or figshare).

@cMadan
Thanks for considering our research work in the Journal of Open Source Software.
I have released a [new version](https://github.com/mir-am/LightTwinSVM/releases) of the program that contains changes from the review process.
I have also uploaded the new version on Zenodo to get a DOI.
[10.5281/zenodo.2619436](https://zenodo.org/record/2619436)
