### 2019-03-31 #1252 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
